Optimised code and changed all values to long long from long

+++ Code/Main.cpp
@@ -5,7 +5,7 @@ using namespace std;
 int main(void) {
     readNames();
     makeGraph();
-    for (vector<long>::iterator i = Graph[4].begin(); i != Graph[4].end(); i++) {
+    for (vector<long long>::iterator i = Graph[4].begin(); i != Graph[4].end(); i++) {
         cout << *i << endl;
     }
     return 0;
